add resolveResult for SimulatorInterface.
